# IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "2.4.0",
    "title" : "CustomerManagement"
  },
  "host" : "serverRoot",
  "basePath" : "/customerManagement",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/customer/{id}/paymentMethod" : {
      "patch" : {
        "tags" : [ "paymentMethod" ],
        "summary" : "Updates paymentMethod of a 'Customer' by Id",
        "operationId" : "patchCustomerPaymentMethod",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "description" : "token to be passed as a header linked to the user whose identity is requested",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Customer",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "PaymentMethodRef",
          "description" : "PaymentMethodRef",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PaymentMethodRef"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/PaymentMethodRef"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "Conflict",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/customer/{id}/contactMedium" : {
      "patch" : {
        "tags" : [ "contactMedium" ],
        "summary" : "Updates contactMedium of a 'Customer' by Id",
        "operationId" : "patchCustomerContactMedium",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "description" : "token to be passed as a header linked to the user whose identity is requested",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Customer",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "ContactMedium",
          "description" : "ContactMedium",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ContactMedium"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/ContactMedium"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "Conflict",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "TimePeriod" : {
      "type" : "object",
      "properties" : {
        "endDateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "An instant of time, ending at the TimePeriod."
        },
        "startDateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "An instant of time, starting at the TimePeriod"
        }
      },
      "description" : "A base / value business entity used to represent a period of time between two timepoints.",
      "example" : {
        "startDateTime" : "2000-01-23T04:56:07.000+00:00",
        "endDateTime" : "2000-01-23T04:56:07.000+00:00"
      }
    },
...
    "PaymentMethodRef" : {
      "type" : "object",
      "required" : [ "href", "id" ],
      "properties" : {
        "href" : {
          "type" : "string",
          "description" : "Reference of the payment mean"
        },
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the payment mean"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the payment mean"
        }
      },
      "description" : "PaymentMethod reference. A payment method defines a specific mean of payment (e.g direct debit).",
      "example" : {
        "name" : "name",
        "href" : "href",
        "id" : "id"
      }
    },
    "ContactMedium" : {
      "type" : "object",
      "required" : [ "characteristic", "contactMediumType" ],
      "properties" : {
        "preferred" : {
          "type" : "boolean",
          "description" : "If true, indicates that is the preferred contact medium"
        },
        "contactMediumType" : {
          "type" : "string",
          "description" : "Type of the contact medium, such as: email address, telephone number, postal address",
          "enum" : [ "emailAddress", "phone", "postalAddress", "fax", "any" ]
        },
        "validFor" : {
          "description" : "The time period that the contact medium is valid for",
          "$ref" : "#/definitions/TimePeriod"
        },
        "characteristic" : {
          "$ref" : "#/definitions/MediumCharacteristic"
        }
      },
      "description" : "Indicates the contact medium that could be used to contact the party.",
      "example" : {
        "contactMediumType" : "emailAddress",
        "validFor" : {
          "startDateTime" : "2000-01-23T04:56:07.000+00:00",
          "endDateTime" : "2000-01-23T04:56:07.000+00:00"
        },
        "preferred" : true,
        "characteristic" : {
          "country" : "country",
          "emailAddress" : "emailAddress",
          "mediumCharacteristicType" : "mediumCharacteristicType",
          "phoneNumber" : "phoneNumber",
          "stateOrProvince" : "stateOrProvince",
          "city" : "city",
          "faxNumber" : "faxNumber",
          "postCode" : "postCode",
          "street1" : "street1",
          "street2" : "street2"
        }
      }
    },
...
    "MediumCharacteristic" : {
      "type" : "object",
      "properties" : {
        "city" : {
          "type" : "string",
          "description" : "The city"
        },
        "country" : {
          "type" : "string",
          "description" : "The country"
        },
        "emailAddress" : {
          "type" : "string",
          "description" : "Full email address in standard format"
        },
        "faxNumber" : {
          "type" : "string",
          "description" : "The fax number of the contact"
        },
        "phoneNumber" : {
          "type" : "string",
          "description" : "The primary phone number of the contact"
        },
        "postCode" : {
          "type" : "string",
          "description" : "Postcode"
        },
        "stateOrProvince" : {
          "type" : "string",
          "description" : "State or province"
        },
        "street1" : {
          "type" : "string",
          "description" : "Describes the street"
        },
        "street2" : {
          "type" : "string",
          "description" : "Complementary street description"
        },
        "mediumCharacteristicType" : {
          "type" : "string",
          "description" : "Type of medium (fax, mobile phone...)"
        }
      },
      "description" : "Describes the contact medium characteristics that could be used to contact a party (an individual or an organization)",
      "example" : {
        "country" : "country",
        "emailAddress" : "emailAddress",
        "mediumCharacteristicType" : "mediumCharacteristicType",
        "phoneNumber" : "phoneNumber",
        "stateOrProvince" : "stateOrProvince",
        "city" : "city",
        "faxNumber" : "faxNumber",
        "postCode" : "postCode",
        "street1" : "street1",
        "street2" : "street2"
      }
    },
...
    "Error" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "description" : "An integer coding the error type. This is given to caller so he can translate them if required."
        },
        "message" : {
          "type" : "string",
          "description" : "A short localized string that describes the error."
        },
        "description" : {
          "type" : "string",
          "description" : "(optional) A long localized error description if needed. It can contain precise information about which parameter is missing, or what are the identifier acceptable values."
        },
        "infoURL" : {
          "type" : "string",
          "description" : "(optional) A URL to online documentation that provides more information about the error."
        }
      }
    }
  }
}